Rename handlers to utils
